Use medium precision in fragment shader and for varyings
Consistently use medium precision for all fragment shader floating
point variables as well as interpolators in interfaces to fragment
shaders. This will have no effect on most desktop hardware, but it
ensures consistent interface definitions on all platforms.

+++ ui/shaders/cellgrid.geom
@@ -13,7 +13,7 @@ in Vertex {
 } v_in[2];
 
 out Varying {
-  smooth float intensity;
+  smooth mediump float intensity;
 } var;
 
 void emitVertexPair(vec4 pos, vec2 shift, float intensity)
